+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+## 0.4.0
+* BREAKING: EventFetchBuilder.fromEvent constructor is replaced by default constructor
+* EventFetchBuilder now internally use EventStreamBuilder, which allow to properly handle initial error
+
 ## 0.3.0
-* New PagedListViewFetcher widget, that fetches a paginated list of data, page by page.
+* New PagedListViewFetcher widget, that fetches a paginated list of data, page by page
 
 ## 0.2.0
 * Add AsyncEditBuilder.fetchingBuilder parameter, to customize the fetching widget independently from the committing widget